Require all DB models have a primary key ID.
We previously allowed a None value to be used to have one generated in
the DAO itself. This makes using the model harder and adds the
temptation of pattern matching out the ID or handling a case in which it
does not exist (even though it will always be present).