Init FB client on another head.
Allows for a faster service boot and does not kill the service entirely
when the FB init step fails. It will later be expanded to add retrying
with a backoff and auto-recovery if the current access token expires.

This removes the dependency on FB and paves the way for improvements to
how external dependencies are addressed in general.